Fixed a few bugs and added new custom icons

+++ Android/app/src/main/java/com/mtspokane/skiapp/activities/activitysummary/ActivitySummary.kt
@@ -52,8 +52,6 @@ class ActivitySummary : FragmentActivity() {
 
 	private lateinit var fileSelectionDialog: FileSelectionDialog
 
-	private lateinit var creditDialog: AlertDialog
-
 	@Deprecated("Fix this")
 	private var mostRecentlyAddedActivityView: ActivityView? = null
 
@@ -113,11 +111,6 @@ class ActivitySummary : FragmentActivity() {
 
 		this.fileSelectionDialog = FileSelectionDialog(this)
 
-		val creditDialogBuilder = AlertDialog.Builder(this)
-		creditDialogBuilder.setView(R.layout.icon_credits)
-		creditDialogBuilder.setPositiveButton(R.string.close_button) { dialog, _ -> dialog.dismiss() }
-		this.creditDialog = creditDialogBuilder.create()
-
 		// Be sure to show the action bar.
 		this.actionBar!!.setDisplayShowTitleEnabled(true)
 
@@ -155,16 +148,7 @@ class ActivitySummary : FragmentActivity() {
 
 				val json: JSONObject = this.getActivitySummaryJson()
 
-				val tmpFileName = "My Skiing Activity.json"
-				this.openFileOutput(tmpFileName, Context.MODE_PRIVATE).use {
-					it.write(json.toString(4).toByteArray())
-				}
-
-				val tmpFile = File(this.filesDir, tmpFileName)
-
-				SkiingActivityManager.shareFile(this, tmpFile, JSON_MIME_TYPE)
-
-				tmpFile.delete()
+				this.writeToShareFile("My Skiing Activity.json", json, JSON_MIME_TYPE)
 			}
 			R.id.share_geojson -> {
 
@@ -172,22 +156,26 @@ class ActivitySummary : FragmentActivity() {
 
 				val geojson: JSONObject = SkiingActivityManager.convertJsonToGeoJson(json)
 
-				val tmpFileName = "My Skiing Activity.geojson"
-				this.openFileOutput(tmpFileName, Context.MODE_PRIVATE).use {
-					it.write(geojson.toString(4).toByteArray())
-				}
-
-				val tmpFile = File(this.filesDir, tmpFileName)
-
-				SkiingActivityManager.shareFile(this, tmpFile, GEOJSON_MIME_TYPE)
-
-				tmpFile.delete()
+				this.writeToShareFile("My Skiing Activity.geojson", geojson, GEOJSON_MIME_TYPE)
 			}
 			R.id.import_activity -> this.importCallback.launch(arrayOf(JSON_MIME_TYPE, GEOJSON_MIME_TYPE))
-			R.id.credits -> this.creditDialog.show()
 		}
 
 		return super.onOptionsItemSelected(item)
+	}
+
+	private fun writeToShareFile(filename: String, jsonToWrite: JSONObject, mime: String) {
+
+		val tmpFile = File(this.filesDir, filename)
+		if (tmpFile.exists()) {
+			tmpFile.delete()
+		}
+
+		this.openFileOutput(filename, Context.MODE_PRIVATE).use {
+			it.write(jsonToWrite.toString(4).toByteArray())
+		}
+
+		SkiingActivityManager.shareFile(this, tmpFile, mime)
 	}
 
 	private fun getActivitySummaryJson(): JSONObject {
@@ -290,7 +278,7 @@ class ActivitySummary : FragmentActivity() {
 		ActivitySummaryLocations.updateLocations(startingActivity)
 
 		// Get the starting activity item.
-		val startingMapMarker: MapMarker = this@ActivitySummary.getActivityItem()
+		val startingMapMarker: MapMarker = this@ActivitySummary.getMapMarker()
 		Log.v(tag, "Starting with ${startingMapMarker.name} (${linkedList.size} items left)")
 
 		// Add the starting circle to the map.
@@ -316,7 +304,7 @@ class ActivitySummary : FragmentActivity() {
 			ActivitySummaryLocations.updateLocations(potentialEndingActivity)
 
 			// Get the ending activity item.
-			val potentialEndingMapMarker: MapMarker = this@ActivitySummary.getActivityItem()
+			val potentialEndingMapMarker: MapMarker = this@ActivitySummary.getMapMarker()
 
 			// Add the circle to the map.
 			this@ActivitySummary.addMapMarkerToMap(potentialEndingMapMarker)
@@ -369,31 +357,32 @@ class ActivitySummary : FragmentActivity() {
 		this@ActivitySummary.addToViewRecursively(linkedList)
 	}
 
-	private fun getActivityItem(): MapMarker {
+	private fun getMapMarker(): MapMarker {
 
-		val chairlift: MapMarker? = ActivitySummaryLocations.checkIfIOnChairlift()
-		val returnPair: MapMarker = if (chairlift != null) {
-			chairlift
-		} else {
-
-			val chairliftTerminal: MapMarker? = ActivitySummaryLocations.checkIfAtChairliftTerminals()
-			if (chairliftTerminal != null) {
-				chairliftTerminal
-			} else {
-
-				val other: MapMarker? = ActivitySummaryLocations.checkIfOnOther()
-				if (other != null) {
-					other
-				} else {
-
-					val run: MapMarker? = ActivitySummaryLocations.checkIfOnRun()
-					run ?: MapMarker(UNKNOWN_LOCATION, ActivitySummaryLocations.currentLocation!!, R.drawable.ic_missing, BitmapDescriptorFactory
-									.defaultMarker(BitmapDescriptorFactory.HUE_MAGENTA), Color.MAGENTA)
-				}
-			}
+		var marker: MapMarker? = ActivitySummaryLocations.checkIfIOnChairlift()
+		if (marker != null) {
+			return marker
 		}
 
-		return returnPair
+		marker = ActivitySummaryLocations.checkIfAtChairliftTerminals()
+		if (marker != null) {
+			return marker
+		}
+
+		marker = ActivitySummaryLocations.checkIfOnOther()
+		if (marker != null) {
+			return marker
+		}
+
+		marker = ActivitySummaryLocations.checkIfOnRun()
+		if (marker != null) {
+			return marker
+		}
+
+		Log.w("getMapMarker", "Unable to determine location")
+		return  MapMarker(UNKNOWN_LOCATION, ActivitySummaryLocations.currentLocation!!,
+			R.drawable.ic_missing, BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_MAGENTA),
+			Color.MAGENTA)
 	}
 
 	@MainThread
